Added locations-suggestions-container component to create-post-page, added event handlers and added / modified effects; add / modified styling properties for locations-suggestions-container component
