+++ client/src/components/locations-suggestions-container/locations-suggestions-container.styles.scss
@@ -1,27 +1,39 @@
 @mixin container {
+  position: absolute;
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
   align-items: center;
   z-index: 20;
+  overflow-y: scroll;
+  background-color: white;
+  border: solid 1px rgb(180, 180, 180);
+  border-radius: 4px;
 }
 
 @mixin element {
-  width: inherit;
+  display: flex;
   background-color: white;
   border-bottom: solid 1px rgb(180, 180, 180);
+  cursor: pointer;
 }
 
 .create-post-locations-suggestions-container {
   @include container();
-  width: 100%;
-  min-width: 250px;
+  min-width: 370px;
+  height: 90px;
+  max-height: 90px;
+  margin-top: -34px;
 
   .create-post-locations-suggestion {
     @include element();
+    width: 352px;
+    min-height: 34px;
+    padding: 4px 8px;
 
     .create-post-location-label {
-      font-size: 18px;
+      font-size: 16px;
+      font-weight: 400;
     }
   }
 }
